+++ src/adjacency_list_weighted/mod.rs
@@ -1,4 +1,4 @@
-//! An adjacency list representation of an arc-weighted digraph.
+//! An adjacency list representation of an arc-weighted digraph
 
 pub mod digraph;
 pub mod fixture;