+++ src/algo/bellman_ford_moore.rs
@@ -1,4 +1,4 @@
-//! The Bellman-Ford-Moore algorithm.
+//! Bellman-Ford-Moore
 //!
 //! The Bellman-Ford-Moore algorithm is a single-source shortest-path algorithm
 //! for weighted digraphs.