+++ src/algo/mod.rs
@@ -4,9 +4,13 @@
 
 pub mod bellman_ford_moore;
 pub mod bfs;
-pub mod breadth_first_tree;
 pub mod dfs;
 pub mod dijkstra;
-pub mod distance_matrix;
 pub mod floyd_warshall;
 pub mod tarjan;
+pub mod types;
+
+pub use types::{
+    breadth_first_tree::BreadthFirstTree,
+    distance_matrix::DistanceMatrix,
+};