+++ src/algo/types/mod.rs
@@ -1,0 +1,4 @@
+//! # Digraph algorithm results
+
+pub mod breadth_first_tree;
+pub mod distance_matrix;